requirements.txt
bcrypt==4.1.2
cryptography==42.0.7
redis==5.0.1
PyYAML==6.0.1
